requirements.txt
# Main dependencies
torch>=2.0.0
transformers>=4.30.0
peft>=0.5.0
datasets>=2.14.0
accelerate>=0.21.0
mlflow>=2.0.0
psycopg2-binary>=2.9.0  # For PostgreSQL backend in production
sqlalchemy>=1.4.0  # For database operations
dvc>=3.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
boto3>=1.26.0
huggingface-hub>=0.16.0
hydra-core>=1.3.0
omegaconf>=2.3.0
scikit-learn>=1.3.0
prometheus-client>=0.16.0
prometheus-fastapi-instrumentator>=6.1.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
ruff>=0.1.0
mypy>=1.7.0
bandit>=1.7.0
pre-commit>=3.0.0
requests>=2.28.0
httpx>=0.24.0
docker>=6.0.0
pyyaml>=6.0.0
